requirements.txt
python-dotenv
requests
google-genai